Add Notion pipeline in NotionEventSyncManager, todo: add manual sync command next
